fix snackbar position for bottom sheets and remove redundent code in home feature

+++ feature/home/presentation/src/main/java/com/sanaa/presentation/bottomsheet/saveToListBottomsheet/SaveToListUiState.kt
@@ -3,19 +3,19 @@ package com.sanaa.presentation.bottomsheet.saveToListBottomsheet
 import com.sanaa.presentation.components.SnackData
 
 data class SaveToListUiState(
-    val mediaId:Long? = null,
+    val mediaId: Int? = null,
     val playlists: List<PlaylistUiItem> = emptyList(),
-    val selectedListsIds: MutableList<Long> = mutableListOf(),
+    val selectedListsIds: MutableList<Int> = mutableListOf(),
     val isLoading: Boolean = false,
-    val isUploading:Boolean = false,
+    val isUploading: Boolean = false,
     val isAddButtonEnabled: Boolean = false,
     val snackBarData: SnackData? = null
 )
 
 data class PlaylistUiItem(
-    val id: Long,
+    val id: Int,
     val title: String,
     val itemCount: Int,
-    val itemsIds: List<Long>,
-    var containsMediaItem:Boolean = false
+    val itemsIds: List<Int>,
+    var containsMediaItem: Boolean = false
 )
