+++ feature/search/presentation/src/main/java/com/sanaa/presentation/bottomsheets/saveToListBottomsheet/SavetoListUiState.kt
@@ -3,9 +3,9 @@ package com.sanaa.presentation.bottomsheets.saveToListBottomsheet
 import com.sanaa.presentation.screen.componants.SnackData
 
 data class SaveToListUiState(
-    val mediaId:Long? = null,
+    val mediaId:Int? = null,
     val playlists: List<PlaylistUiItem> = emptyList(),
-    val selectedListsIds: MutableList<Long> = mutableListOf(),
+    val selectedListsIds: MutableList<Int> = mutableListOf(),
     val isLoading: Boolean = false,
     val isUploading:Boolean = false,
     val isAddButtonEnabled: Boolean = false,
@@ -13,9 +13,9 @@ data class SaveToListUiState(
 )
 
 data class PlaylistUiItem(
-    val id: Long,
+    val id: Int,
     val title: String,
     val itemCount: Int,
-    val itemsIds: List<Long>,
+    val itemsIds: List<Int>,
     var containsMediaItem:Boolean = false
 )
